backend/requirements.txt
django>=5.2.8
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=22.0
whitenoise>=6.7
psycopg2-binary>=2.9
dj-database-url>=2.1
